ref(backend>user): Removed plural + Moved user schema from login to user
