+++ backend/src/app.module.ts
@@ -3,7 +3,7 @@ import { ConfigModule } from '@nestjs/config';
 import { MongooseModule } from '@nestjs/mongoose';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UsersModule } from './users/users.module';
+import { UserModule } from './user/user.module';
 import { AuthModule } from './auth/auth.module';
 import { CommonModule } from './common/common.module';
 import { DatabaseModule } from './database/database.module';
@@ -13,7 +13,7 @@ import { LoginModule } from './login/login.module';
   imports: [
     MongooseModule.forRoot('mongodb://mongo:27017/nestdb'),
     ConfigModule.forRoot({ isGlobal: true }),
-    UsersModule,
+    UserModule,
     AuthModule,
     CommonModule,
     DatabaseModule,
