--- backend/src/login/schemas/user.schema.ts
@@ -1,16 +1,0 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { Document } from 'mongoose';
-
-@Schema()
-export class User extends Document {
-  @Prop({ required: true, unique: true })
-  email!: string;
-
-  @Prop({ required: true })
-  password!: string;
-
-  @Prop({ required: true })
-  name!: string;
-}
-
-export const UserSchema = SchemaFactory.createForClass(User);
