+++ backend/src/user/userController.ts
@@ -1,4 +1,4 @@
 import { Controller } from '@nestjs/common';
 
 @Controller('users')
-export class UsersController {}
+export class UserController {}
